fix: do not clean up unused credentials if current set is invalid

+++ controllers/azureadapplication/azureadapplication_controller.go
@@ -121,7 +121,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 		return r.HandleError(*tx, err)
 	}
 
-	if tx.Options.Process.Secret.Cleanup {
+	if tx.Options.Process.Secret.Cleanup && tx.Options.Process.Secret.Valid {
 		err = r.Azure().DeleteUnusedCredentials(*tx)
 		if err != nil {
 			return r.HandleError(*tx, err)
